Fix context trail counter justification polarity and show root target's relation to context tail
